removing DBTableList variable from knex queries

+++ src/services/store/store.service.ts
@@ -1,5 +1,5 @@
 import config from '#src/config/index.js'
-import db, { DBTableList } from '#src/lib/knex/db.js'
+import db from '#src/lib/knex/db.js'
 import { Hearts } from '#src/models/hearts.js'
 import { Review } from '#src/models/review.model.js'
 import { Store } from '#src/models/store.model.js'
@@ -73,14 +73,14 @@ export async function getStores(opt: { page: number }) {
 }
 
 export async function getStoreCount() {
-    return db.table<Store>(DBTableList.STORE_TABLE).count({ count: '*' })
+    return db.table<Store>('stores').count({ count: '*' })
 }
 
 export async function getStoreBySlug(slug: string) {
     return db
         .with('stores', (qb) => {
-            qb.table<Store>(DBTableList.STORE_TABLE)
-                .select(`${DBTableList.STORE_TABLE}.*`)
+            qb.table<Store>('stores')
+                .select(`${'stores'}.*`)
                 .where('slug', slug)
                 .modify(joinTagsStores)
         })
@@ -91,7 +91,7 @@ export async function getStoreBySlug(slug: string) {
 
 export async function countBySlug(slug: string, exceptStoreId?: number) {
     return db
-        .table<Store>(DBTableList.STORE_TABLE)
+        .table<Store>('stores')
         .count('slug', { as: 'slug' })
         .where('slug', slug)
         .modify((qb) => {
@@ -103,7 +103,7 @@ export async function countBySlug(slug: string, exceptStoreId?: number) {
 
 export async function getStoreByTag(tag: string) {
     return db
-        .table<Store>(DBTableList.STORE_TABLE)
+        .table<Store>('stores')
         .select('stores.*')
         .modify(joinTagsStores)
         .modify((qb) => {
@@ -115,34 +115,26 @@ export async function getStoreByTag(tag: string) {
 
 export async function getHeartsOfUser(user_id: number) {
     return db
-        .table<Hearts>(DBTableList.HEART_TABLE)
-        .select(`${DBTableList.STORE_TABLE}.*`)
+        .table<Hearts>('hearts')
+        .select(`stores.*`)
         .where('user_id', user_id)
-        .innerJoin(DBTableList.STORE_TABLE, (qb) => {
-            qb.on(
-                `${DBTableList.STORE_TABLE}.id`,
-                '=',
-                `${DBTableList.HEART_TABLE}.store_id`
-            )
+        .innerJoin('stores', (qb) => {
+            qb.on(`stores.id`, '=', `hearts.store_id`)
         })
         .modify(joinTagsStores)
 }
 
 export async function toggleHeartStore(store_id: number, user_id: number) {
     const isHearted = await db
-        .table<Hearts>(DBTableList.HEART_TABLE)
+        .table<Hearts>('hearts')
         .where({ store_id, user_id })
         .first()
+
     if (isHearted) {
-        return db
-            .table<Hearts>(DBTableList.HEART_TABLE)
-            .where({ store_id, user_id })
-            .delete()
+        return db.table<Hearts>('hearts').where({ store_id, user_id }).delete()
     }
 
-    return db
-        .table<Hearts>(DBTableList.HEART_TABLE)
-        .insert({ store_id, user_id })
+    return db.table<Hearts>('hearts').insert({ store_id, user_id })
 }
 
 export async function getTags() {
@@ -161,7 +153,7 @@ export async function getTags() {
 
 export async function getStoreNearPoint(x: number, y: number, distance = 1000) {
     return db
-        .table<Store>(DBTableList.STORE_TABLE)
+        .table<Store>('stores')
         .select(
             'id',
             'slug',
@@ -180,19 +172,11 @@ export async function getStoreNearPoint(x: number, y: number, distance = 1000) {
 
 export async function getTopStores() {
     return db
-        .table<Store>(DBTableList.STORE_TABLE)
-        .innerJoin<Review>(DBTableList.REVIEW_TABLE, (qb) => {
-            qb.on(
-                db.raw(
-                    `${DBTableList.STORE_TABLE}.id = ${DBTableList.REVIEW_TABLE}.store`
-                )
-            )
+        .table<Store>('stores')
+        .innerJoin<Review>('reviews', (qb) => {
+            qb.on(db.raw(`stores.id = reviews.store`))
         })
-        .select(
-            db.raw(
-                `${DBTableList.STORE_TABLE}.*, AVG(reviews.rating) as avg_rating`
-            )
-        )
+        .select(db.raw(`stores.*, AVG(reviews.rating) as avg_rating`))
         .groupBy('stores.id')
         .having(db.raw('COUNT(reviews.id) > 1'))
         .orderBy('avg_rating', 'desc')
